Repository class명 변경, api_client를 통해 api 통신
