+++ lib/presentation/keyword_cloud_screen/controller/keyword_cloud_controller.dart
@@ -1,4 +1,3 @@
-import '../../../data/repository/post_repository.dart';
 import '/core/app_export.dart';
 import 'package:application4/presentation/keyword_cloud_screen/models/keyword_cloud_model.dart';
 
@@ -8,7 +7,7 @@ class KeywordCloudController extends GetxController {
   var userid;
   var keywords = <String>[].obs;
   int maxKeywords = 10;
-  MyRepository myRepository = Get.find<MyRepository>();
+  Repository myRepository = Get.find<Repository>();
 
   @override
   void onInit() async{
@@ -16,12 +15,15 @@ class KeywordCloudController extends GetxController {
     userid = myRepository.userid;
   }
 
+  /// keyword 추가, 삭제
   onoffKeyword(String keyword){
+    // 키워드가 선택되어 있는 상태에서 클릭하면 키워드를 지움
     if (keywords.contains(keyword)) {
       deleteKeyword(keyword);
       update();
       return false;
     }
+    // 키워드가 선택되어 있지 않은 상태에서 클릭하면 키워드를 추가함
     else {
       if(keywords.length >= maxKeywords) return false;
       addKeyword(keyword);
@@ -40,6 +42,7 @@ class KeywordCloudController extends GetxController {
 
   /// keyword 제출
   submitKeyword() async{
+    // 선택한 키워드를 repository, apiclient를 통해 서버로 전달
     var data = List.generate(0, (index) => Map<String, String>());
     keywords.forEach((element) {
       data.add({
@@ -48,6 +51,7 @@ class KeywordCloudController extends GetxController {
       });
     });
 
+    // api가 성공했는지 체크
     bool isposted = await myRepository.postKeyword(data);
     if (isposted) {
       print("keyword is posted.\n");
