+++ lib/presentation/signup_screen/controller/signup_controller.dart
@@ -2,7 +2,6 @@ import 'package:application4/core/app_export.dart';
 import 'package:application4/presentation/signup_screen/models/signup_model.dart';
 
 import '../../../data/controllers/user_controller.dart';
-import '../../../data/repository/post_repository.dart';
 
 class SignupController extends GetxController {
   Rx<SignupModel> signupModelObj = SignupModel().obs;
@@ -12,9 +11,11 @@ class SignupController extends GetxController {
   // 0:중복확인버튼아직안누름, 1:id중복, 2:통과(id중복되지않음)
   RxInt isduplication = 0.obs;
 
-  MyRepository myRepository = Get.find<MyRepository>();
+  Repository myRepository = Get.find<Repository>();
   var userController = Get.find<UserController>();
 
+
+  // 회원정보 목록 및 선택 함수
 
   List<String> dropdownListAge = ['10대', '20대', '30대', '40대', '50대', '60대 이상'];
   String selectedDropdownAge = '10대';
@@ -40,6 +41,9 @@ class SignupController extends GetxController {
     update();
   }
 
+
+
+  /// 중복인지 확인하는 함수. 결과를 signup controller의 isduplication에 저장함
   isDuplication() async {
     bool isidexist = await userController.isIdExist("0");
     if (!isidexist)
@@ -48,12 +52,13 @@ class SignupController extends GetxController {
       this.isduplication.value = 1;
   }
 
-  ///userid 등록 함수
+  /// userid 등록 함수
   setID(var userid) {
     this.userid = userid;
     this.userController.setID(userid);
   }
 
+  /// 회원가입 정보를 서버에 등록하는 함수
   postUserData() async {
     var data = {
     "ID": "${this.userid}",
